Change what in layer_state_t to 64 bit.
We finally reached the edge of 32 bit enum for it in ag/5659622.

Bug: None
Test: Builds.

+++ libs/gui/include/gui/LayerState.h
@@ -124,7 +124,7 @@ struct layer_state_t {
         float dsdy{0};
     };
     sp<IBinder> surface;
-    uint32_t what;
+    uint64_t what;
     float x;
     float y;
     int32_t z;
